+++ src/index.ts
@@ -1,12 +1,14 @@
 import {default as HTMLIntlDisplayNamesElement} from './elements/displaynames/displaynames';
 import {default as HTMLIntlListFormatElement} from './elements/listformat/listformat';
 import {default as HTMLIntlListItemElement} from './elements/listitem/listitem';
+import {default as HTMLIntlRelativeTimeFormatElement} from './elements/relativetimeformat/relativetimeformat';
 
 declare global {
   interface HTMLElementTagNameMap {
     'intl-displaynames': HTMLIntlDisplayNamesElement;
     'intl-listformat': HTMLIntlListFormatElement;
     'intl-listitem': HTMLIntlListItemElement;
+    'intl-relativetimeformat': HTMLIntlRelativeTimeFormatElement;
   }
 }
 
@@ -22,14 +24,20 @@ export function defineIntlListItemElement() {
   customElements.define('intl-listitem', HTMLIntlListItemElement);
 }
 
+export function defineIntlRelativeTimeFormatElement() {
+  customElements.define('intl-relativetimeformat', HTMLIntlRelativeTimeFormatElement);
+}
+
 export function defineIntlElements() {
   defineIntlListItemElement();
   defineIntlDisplayNamesElement();
   defineIntlListFormatElement();
+  defineIntlRelativeTimeFormatElement();
 }
 
 export {
   HTMLIntlDisplayNamesElement,
   HTMLIntlListFormatElement,
   HTMLIntlListItemElement,
+  HTMLIntlRelativeTimeFormatElement,
 }