+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - [`intl-listformat`]: Added `formatToParts()` method
+- [`intl-relativetimeformat`] Added the custom element
 
 ## [0.0.1-alpha.7] - 2022-11-19